merge: dev/mid 머지 전 수정

+++ server/src/main/java/com/asap/server/repository/CodeBattleMatchRepository.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
 import com.asap.server.domain.CodeBattleMatch;
 
@@ -27,16 +27,16 @@ public interface CodeBattleMatchRepository extends JpaRepository<CodeBattleMatch
             Long contestId,
             Long user1Id,
             Long user2Id);
-            
+
     List<CodeBattleMatch> findByContestIdAndUser1Id(
-        Long contestId,
-        Long user1Id);
+            Long contestId,
+            Long user1Id);
 
     long countByContestIdAndLogIsNotNull(Long contestId);
 
     @Query("SELECT COUNT(m) FROM CodeBattleMatch m WHERE m.contest.id = :contestId AND m.log IS NOT NULL")
     long countFinishedMatchesByContestId(@Param("contestId") Long contestId);
 
-    Optional<CodeBattleMatch> findBySubmissionIdAndAiOrder(Long submissionId, Integer aiOrder);
+    Optional<CodeBattleMatch> findBySubmissionIdAndAiOrder(Long submissionId, Long aiOrder);
 
 }
